+++ example/lib/events/todo_added_event.dart
@@ -1,0 +1,6 @@
+import 'package:example/sdk/todo/proxies/todo.dart';
+
+class TodoAddedEvent {
+  final Todo todo;
+  TodoAddedEvent(this.todo);
+}