+++ example/lib/events/todo_updated_event.dart
@@ -1,0 +1,6 @@
+import 'package:example/sdk/todo/proxies/todo.dart';
+
+class TodoUpdatedEvent {
+  final Todo todo;
+  TodoUpdatedEvent(this.todo);
+}