+++ CHANGELOG.md
@@ -15,3 +15,7 @@
 * Safe state disposal. Ignore `triggerStateChange()` when the state is disposed
 * Update StorageService and added the ability to check if key exists `contains` 
 * `KeepAliveClientWidgetStateBase` for clients of AutomaticKeepAlive (example: ListView). This keeps the state of a widget alive, given the wantAlive is set to true.
+
+## [0.0.4] - October 27, 2020
+* Event Aggregator (automatically registered as a singleton)
+* BREAKING CHANGE: The StorageService renamed to SecureStorageService, and is registered as a singleton automatically if not registered by user.